bonusPointsRender 파일 분리
