+++ src/basic/utils/dateUtils.js
@@ -1,0 +1,3 @@
+export function isTuesday() {
+  return new Date().getDay() === 2;
+}
